Chunk rendering update

+++ Client/SDL_project/SpaceGame.h
@@ -84,7 +84,7 @@ public:
 	//! Coordinates of the mouse 
 	int mouse_X, mouse_Y;
 
-	int cameraX = 0, cameraY = 0;
+	float cameraX = 0, cameraY = 0;
 
 	bool spawnPlayer = true;
